Bugfix with closing a busy, waiting connection.
Be sure to process reserved jobs after removing the connection from
any list it is in, especially the waiting list.

There was a problem while closing a connection that both has reserved
jobs and is waiting for a job. Processing the reserved jobs before
removing the closing connection from the wait list (i.e. in the wrong
order) was causing us to try to assign a reserved job to the same
connection that is in the process of being closed, possibly failing to
assign a job to another valid connection, closing this connection
twice, and corrupting the stats.