A bunch more stats and a version header.
Almost all stats from memcached that make sense in beanstalkd have
been added.

The packaging code no longer picks a version number automatically, you
must specify when you run make pkg.

There is now a version.h that defines the version string. It will
always be "dev" in the source tree, but that file is replaced in
release packages with the release version.