Report errors just like all other messages.
One of the principles of the beanstalk protocol is that messages
should be readable by both machines and humans. The old error messages
didn't satisfy that property -- the numeric error codes and
human-readable descriptions were distinct.

So instead we'll deliver errors the same way as other messages: as a
single descriptive upper-case word, possibly with underscores.