Only update the memo if it already exists.
Be more conservative when updating the memo. If the memo is set, then it's
guaranteed to be the soonest job and we must preserve this invariant. But if
the memo is null, then we don't know anything about the reserved jobs; we can't
assume that job j is the soonest.