Add (experimental) binlog support for persistence
Binary log files are created in a directory. Each time a job
changes status a record will be written to the file.

Each job keeps a reference count on the binlog it was created in.
Logs are purged when their reference count hits zero and all
binlogs prior have also hit zero

On server startup the binary logs are replayed to create a queue
with all the jobs either in ready, delayed or buried state. Basically
similar to what it would look like if all clients just disconnected