Only create a new job from binlog when we have a tube name
A READY or DELAYED entry in the binlog only has a tubename when the
job is first created. A release of a job with a delay will insert a new
entry without a tubename.

This fixes a problem when on replay of the binlog a record is
encountered that is modifying an existing job. But that job was originally
created in a prior binlog file that has been deleted becasue all
its jobs have been deleted. ie a modification to a job that has
already completed so we do not need to create it