Reserve space for certain events ahead of time.
Delete records are appended to the binlog just like everything else.
That means if there's no disk space available, the user won't be able to
delete jobs in an attempt to free up disk space. So we allocate and
reserve space ahead of time for deleting a job.

Also quite a bit of refactoring in this commit.