Only read a job body for new jobs
The body of a job is only written to the binlog when a job is
first created. all other times only the job struct is written.

For new jobs the tube namelen will be >0, if this is an existing
job the tube_namelen will be zero, so do not attemtp to read the
job body