Integrate srvschedconn() into connsched().
The tickat value of the connection cannot be modified before removing it from a
heap.  This would compromise the heap ordering.  Therefore, the srvschedconn()
function needs to be split so the tickat value can be updated.

As the srvschedconn() is not used anywhere else apart from connsched(), it can
be removed.