remove markdown version of protocol.txt
It's completely redundant, and the benefit in
readability (if there is any) isn't worth the effort
keeping them in sync.

The Chinese tranlation is not redundant, so we should
keep it. But still, note that it's not authoritative.
Where protocol.txt and protocol.zh-CN.md disagree,
protocol.txt is correct.

Fixes #308
Fixes #191