chore(adm): remove all init files
Proposal here is to have this repository not focus on surrounding
infrastructure needed to run it in various environments; this means
_not_ maintaining

 * init system configuration files.
 * packaging stuff (RPM, Deb, Docker etc.).

**Why?** Two reasons:

 * Because it doesn't scale. There are lots of different system
   setups.
 * Because this helps us stay focused on the problem we are trying to
   solve, which is to have a badass and super useful small queue.

**So, where do we put packaging and system init configuration files?**
Let's maintain that in a separate repositories as long as someone can
take ownership of it. Just create a pull request and ping me and I'll
create the repo.

This commit makes PR #178 & #175 obsolete.