set input width of pri, delay and ttr params to uint32
Before this fix if the value specified in
pri/delay/ttr parameters would equal to 2**32
then it would be messed up and set to zero because of overflow.
And in general, the behaviour was not defined strictly
because the uint type can take different width.

I have fixed the read_num function to parse integer
that fits into range 0 .. 2**32-1. Internally it parses
integer into uint64, but it checks if it fits into uint32.
So it will return success only in this case.

I have added multiple tests checking that behaviour.

Fixes #230