fix coverage for forked processes
Previously, the forked process with a server was killed by SIGKILL.
This prevented the process from writing coverage data.

This patch enables the child processes to handle the SIGUSR2 signal.
testserv.c setups a handler that flushes coverage data prior exiting.
This handler is setup only when WITHCOV is defined in CFLAGS.

Sneakily move code quality badges to the top of README.

Coverage after the patch:

lines: 64.9% (1434 out of 2209)
branches: 50.9% (649 out of 1274)

Updates #424
Fixes #433