use size_t for indexes in the heap structure
Previously "int" was used, it may be too small on some architectures.
This change uses "size_t" instead, since index cannot be negative.

In the heapremove function it removes the assignment to the heap_index
field of deleted member.

As usual in affected functions, I have moved declarations of variables close
to their usage.

Updates #443