fix the job size overflowing the int type
I have set the maximum allowed limit for the job size to 1024^3 bytes (1GB).
When user tries to specify (-z) the limit for the job size larger than that,
he will get a warning that limit was capped to 1GB.

The code stores body_size in int32 and verifies that parameter in the
put command can be parsed. I have propagated the change to the type
to all involved functions.

Fixes #401