include dependent headers in the dat.h file (#460)
Before this fix Emacs was hinting the *_t types as undefined. This was confusing.
Include the required headers in the dat.h file.

I have rearranged dat.h and sd-daemon.h before other headers to verify this.
I have removed redefinitions of *_t types because it was producing redefinitions errors.