fix the close_releases_job test
On some machines the release might not happen soon enough
after the connection holding the job was closed. Instead of
adding artifical timeout I have added reserving with timeout of 1 second.
Test will pass if the job is released sooner than in 1 second.
It should be enought even for the most slow machines.

Fixes #475