fix prototypes of the Less and Record typedefs
This change improves readability of the code.
It brings consistency into functions of the binary heap.
Typedef Less was renamed to less_fn, Record to setpos_fn.
connless and connrec were changed to match to the corresponding typdefs.

The rec member of heap was renamed to "setpos".

The job was updated in the same way.

Fixes #472