fix benchmarks in testsrv.c
The slowness in benchmarks on Linux was due to long wait time
in the select syscall. Setting socket's TCP_NODELAY helped.

Also see
https://stackoverflow.com/questions/39270419/select-on-socket-slow-in-linux

I added proper resettimers inside of benchmarks to skip the overhead.
This helped to improve the time spent in all of testsrv tests too.

Bechmarking is not ideal yet. It might be capped by the client's limitations.

Fixes #430