handle SIGTERM to allow quick exit under Docker
Before this patch: when beanstalkd with PID=1 received SIGTERM
running inside of Docker container, the process never exited.
Docker SIGKILLed the process after 10s timeout.

This change adds an ability to handle such case without timeout.
Server will signal itself SIGKILL when SIGTERM is received.
This allows instant termination. This is part of the crash-only
strategy so crashing this way is okay.

Fixes #527