support UNIX domain sockets
This change adds support for listening to UNIX sockets.
It extends the semantics of the "-l" command line option to support creating local
sockets by using the "unix:" prefix, similarly to nginx.

It updates the man page to explain local socket support.
It adds some tests over unix sockets.

It uses a generic address storage for accept().
Now that AF_UNIX socket family is supported, a more generic sockaddr variant can
be used.  This is not essential for local socket support, it is only a way to be
more pedantic.