disable forced binding to IPV6-only stack if "-l" was not set
This issue appears on FreeBSD 8.4 where IPV6_V6ONLY is set to 1 by
default. This patch unsets this flag if host was not specified.

Each OS choses the default value for this flag, but it's better
to override it for the case when listening on all interface is
expected.

Fixes #228