add the "reserve-job" command
A job can be reserved by an ID in the Ready, Buried or Delayed states.
Reservation made with this command does not differ from reservations
made by "reserve" or "reserve-with-timeout" commands. The new command
does not produce "deadline soon" when there are other jobs about to expire.

Fixes #310
Fixes #222
Fixes #240